refactor(app): clean up achievement component structure
