+++ src/app/page.tsx
@@ -1,6 +1,5 @@
-import HeroLandingContainer from '@/components/pages/home/Hero/HeroLandingContainer'
-import AchievementsLandingContainer from '@/components/pages/home/achievement/achievementLandingContainer'
-import CallToActionLandingContainer from '@/components/pages/home/callToAction/callToActionLandingContainer'
+import AchievementsLandingContainer from '@/components/pages/home/achievement'
+import HeroLandingContainer from '@/components/pages/home/hero'
 import MissionLandingContainer from '@/components/pages/home/mission/missionLandingContainer'
 
 export default function Home() {
@@ -9,7 +8,6 @@ export default function Home() {
       <HeroLandingContainer />
       <MissionLandingContainer />
       <AchievementsLandingContainer />
-      <CallToActionLandingContainer />
     </div>
   )
 }
