--- src/components/pages/contact/Hero/index.tsx
@@ -1,5 +1,0 @@
-const HeroSection = () => {
-  return <div>HeroSection</div>
-}
-
-export default HeroSection
